analysis/requirements.txt
Flask
functions-framework==3.*
numpy
pandas
pandas-gbq
matplotlib
seaborn
google-cloud-storage
